requirements.txt
streamlit
python-docx
pywin32-ctypes
pandas
numpy
pyarrow
Pillow